This commit was SVN r2876.

The following SVN revision numbers were found above:
  r2807 --> 2921a975bbd2e52b37c3a9432b2746eec60480ab
  r2875 --> 65a5bfa3dc47af101bf4f0df6423add18955a254

Revision number ranges (suitable for "git log"):
  r2807:2875 --> 2921a975..65a5bfa3

+++ utils/hwloc-distrib.c
@@ -196,7 +196,7 @@ int main(int argc, char *argv[])
 
       hwloc_distributev(topology, roots, chunks, cpuset, n, to_depth);
 
-      for (i = 0; i < n; i++) {
+      for (i = 0; (long) i < n; i++) {
 	char *str = NULL;
 	if (singlify)
 	  hwloc_bitmap_singlify(cpuset[i]);
